Update readme
Updated readme to note needed dependency of Net::DNS CPAN package libidn.
Also noted that sudo is needed to run the script to be able to bind to the port for a DNS server.

Altered list of steps needed to use the script to be bullet points to avoid glitchiness of github numbered lists restarting.